// src/commons/seguimirntoPlanes/SeguiminetoPOAI/components/Components/BusquedaAvanzada/types.ts
export interface IBusquedaSeguiminetoPOAI {
    id: number;
    id_seguimiento: number;
    nombre_plan: string;
    nombre_programa: string;
    nombre_proyecto: string;
    nombre_producto: string;
    nombre_actividad: string;
    nombre_unidad: string;
    nombre_indicador: string;
    nombre_meta: string;
    codigo_modalidad: string;
    concepto: string;
    sector: string;
    nombre_fuente: string;
    cuenta: string;
    objeto_contrato: string;
    ubicacion: string;
    clase_tercero: string;
    porcentaje_pto: number;
    vano_1: number;
    vano_2: number;
    vano_3: number;
    vano_4: number;
    valor_total: number;
    numero_cdp_paa: number;
    numero_rp_paa: number;
    valor_seguimiento_banco_paa: number;
    valor_cdp_paa: number;
    valor_rp_paa: number;
    fecha_termiacion: string;
    duracion: number;
    valor_mesual_paoi: number;
    mes_oferta_paa: string;
    numero_cdp: number;
    mes_solicita: string;
    valor_pagado: number;
    valor_obligado: number;
    valor_saldo: number;
    porcentaje_ejecuta: number;
    numero_contrato: number;
    numerp_rp: number;
    fecha_rp: string;
    valor_cdp: number;
    fecha_cdp: string;
    observaciones: string;
    id_plan: number;
    id_programa: number;
    id_proyecto: number;
    id_producto: number;
    id_actividad: number;
    id_indicador: number;
    id_meta: number;
    id_concepto: number;
    id_fuente_financiacion: number;
    id_unidad_organizacional: number;
    id_detalle_inversion: number;
    id_banco_proyecto: number;
    id_modalidad: number;
    id_ubicacion: number;
    id_clase_tercero: number;
    id_sector: number;
}